requirements.txt
arxiv
streamlit>=1.28.0
langchain
langchain-chroma
langchain-core
langchain-google-genai
chromadb
pydantic
python-dotenv
requests
